+++ Source/Renderer/Source/Texture.cpp
@@ -63,7 +63,7 @@ void Texture::Release()
 {
 	if (_srvArray.size() > 0)
 	{
-		for (int i = 0; i < _srvArray.size(); ++i)
+		for (size_t i = 0; i < _srvArray.size(); ++i)
 		{
 			if(_srvArray[i])
 			{
